mySecret/requirements.txt
python-dotenv==0.15.0
cryptography==3.2.1